+++ django/projects/mysite/settings_base.py
@@ -238,6 +238,7 @@ MEDIA_EXPORT_SUBDIRECTORY = 'export'
 # Cropping output extension
 CROPPING_OUTPUT_FILE_EXTENSION = "tiff"
 CROPPING_OUTPUT_FILE_PREFIX = "crop_"
+CROPPING_VERIFY_CERTIFICATES = True
 
 # The maximum allowed size in Bytes for generated files. The cropping tool, for
 # instance, uses this to cancel a request if the generated file grows larger